Change modman module to use a structure enabling real builtin
This changes libmodman internal module ABI/API to use a single structure
to store module information. We now have a single structure to load from
the shared object. This same structure is being passed to load builtin
module in a way that does not rely uppon dlopen/dlsym. See issue 133.